+++ apps/desktop/src/preload.js
@@ -1,0 +1,7 @@
+const { contextBridge, ipcRenderer } = require('electron');
+
+contextBridge.exposeInMainWorld('eduvision', {
+  close: () => ipcRenderer.send('app-close'),
+  minimize: () => ipcRenderer.send('app-minimize'),
+  maximize: () => ipcRenderer.send('app-maximize'),
+});